requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.1.0
anthropic==0.52.0
stripe==12.1.0
httpx==0.28.1
openpyxl==3.1.5
openregister-sdk @ git+https://github.com/oregister/openregister-python.git
pydantic>=2.0.0
svix>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
